Made changes for Qt6 compatability
Unfortunately task bar progress was lost due to the removal of the
winextras module from Qt6. It is noted in the Qt docs it may return. If
it doesn't, look into native solution.

+++ src/flashpoint.cpp
@@ -114,7 +114,7 @@ bool operator== (const AddApp& lhs, const AddApp& rhs) noexcept
 }
 
 //-Hashing------------------------------------------------------------------------------------------------------
-uint qHash(const AddApp& key, uint seed) noexcept
+size_t qHash(const AddApp& key, size_t seed) noexcept
 {
     QtPrivate::QHashCombine hash;
     seed = hash(seed, key.mID);
