+++ src/flashpoint.h
@@ -137,7 +137,7 @@ public:
 
 //-Hashing-------------------------------------------------------------------------------------------------------------
 public:
-    friend uint qHash(const AddApp& key, uint seed) noexcept;
+    friend size_t qHash(const AddApp& key, size_t seed) noexcept;
 
 //-Instance Functions------------------------------------------------------------------------------------------------------
 public:
